Added static_cast to remove LSP-error messages.
The implicit narrowing convertion is allowed by the compiler settings
for Arduino NANO33. While technically the conversion in this case is
narrowing, we do not lose any information as we are explicitly
packing 16-bits into two 8-bit variables.

It has to be done this way since the USB-descriptors only deal with 8-bit fields.

I have added static_cast<uint8_t> for the 16-bit values before assignment to keep the LSP-happy.